added test files for tb

+++ driver/driver.py
@@ -5,7 +5,7 @@ from pynq import Overlay
 import time
 
 print("Running PYNQ Driver")
-test_data = np.loadtxt('./driver_content/test_data/test_data.txt')
+test_data = np.loadtxt('./driver_content/test_data/test_data.txt', dtype=np.float32)
 test_label_one_hot = np.loadtxt('./driver_content/test_data/test_label_one_hot.txt')
 test_label = np.loadtxt('./driver_content/test_data/test_label.txt', dtype=np.int8)
 print("Test data loaded")
@@ -16,8 +16,8 @@ class Driver:
     def __init__(self):
         self.ol = Overlay(BIT_PATH)
         self.dma = self.ol.axi_dma_0
-        self.input_buffer = allocate(shape=(240,), dtype=np.int32)
-        self.output_buffer = allocate(shape=(6,), dtype=np.int32)
+        self.input_buffer = allocate(shape=(240,), dtype=np.float32)
+        self.output_buffer = allocate(shape=(6,), dtype=np.float32)
     
     def predict(self, x):
         #quantise input
